eth_dmx: my previous assumption was wrong
From the first tests I assumed that the brightness needs to follow the
desired brightness of the fixture (if light should come out, brightness
needs to be 100%, if it should stay dark brightness needs to be 0%).

Turns out, setting the brightness to 100% at boot is fine, and the
remaining brightness control can be done perfectly using the RGB
channels.